Incorrect metadata for sql_variant datatype on a restored server (#131)
When pg_dump/restore is performed on database with table containing sql_variant data type columns, it was observed that the metadata seemed to get lost during the dump process. Therefore, the values were not correctly restored, which reflected in the failure of some test cases in JDBC framework. This change aims to separately handle table containing sql_variant data type columns so that the values can be correctly dumped and restored.

Tests added in extensions PR: babelfish-for-postgresql/babelfish_extensions#1385

Issues Resolved
BABEL-4036
Signed-off-by: Aditya Verma adiityav@amazon.com