Change order of hook call (#142)
During UNION parsing, postgres checks that the length of the top-level
targetlist doesn't change. However, we modify this target list in
pre_transform_sort_clause_hook, and base the original length off the
modified version. The length shouldn't be changing in the hook, but for
safety, we should wait to revert these changes until after this check.

Taks: BABEL-4169
Signed-off-by: Walt Boettge <wboettge@amazon.com>